// web_assets/kone-design-system-main/packages/kone-ds-webcomponents/src/components/molecules/kfds-snackbar/kfds-snackbar.enum.ts
export enum SnackbarVerticalPosition {
    Top = 'top',
    Bottom = 'bottom'
};

export enum SnackbarHorizontalPosition {
    Left = 'left',
    Right = 'right',
    Center = 'center'
};

export enum SnackbarType {
    Info = 'info',
    Warning = 'warning',
    Error = 'error',
    Success = 'success'
};